Promovido los cambios
Terminado CRUD Noticias, Terminado Crud Usuario, Terminado Sesion de Noticias en el Web Site, Creado todos los middleware que se usara en el sistema,Instalado Paquete Flash para el manejo de mensajes de exito o error, el rol postulante se dividio en dos: postulante_becario y postulante_mentor

+++ app/Http/Controllers/MantenimientoNoticiaController.php
@@ -6,6 +6,9 @@ use avaa\Editor;
 use avaa\Noticia;
 use Illuminate\Http\Request;
 use avaa\Http\Controllers\Controller;
+use Redirect;
+use Yajra\Datatables\Datatables;
+use Laracasts\Flash\Flash;
 
 class MantenimientoNoticiaController extends Controller
 {
@@ -14,8 +17,14 @@ class MantenimientoNoticiaController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('editor');
+    }
     public function index(Request $request)
     {
+        /* Esto se comento porque ahora se usa AJAX
         $noticias = Noticia::search($request->titulo)->orderBy('id','ASC')->paginate(4);
 
         //each lo que hace es un recorrido por cada uno de las noticias
@@ -23,11 +32,13 @@ class MantenimientoNoticiaController extends Controller
             $noticias->editor->user; //aqui lo que hago es llamar a la relaciones para cada uno de las noticias
 
         });
+        */
 
 
 
-        return view('sisbeca.crudNoticia.mantenimientoNoticia')->with('noticias',$noticias)->with('titulo',$request->titulo);
+        return view('sisbeca.crudNoticia.mantenimientoNoticia');
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -72,9 +83,18 @@ class MantenimientoNoticiaController extends Controller
 
             $noticia->url_imagen = '/images/noticias/'.$name;
 
-            $noticia->save();
+            if($noticia->save())
+            {
+                flash('Articulo Registrado Exitosamente!','success')->important();
+            }
+            else
+            {
+                flash('Ha ocurrido un error al registrar el articulo')->error()->important();
+            }
 
 
+        }else{
+            flash('Debe Ingresar una imagen para registrar el articulo')->error()->important();
         }
 
 
@@ -106,7 +126,9 @@ class MantenimientoNoticiaController extends Controller
 
         if(is_null($noticia))
         {
-            abort('404','Archivo no encontrado');
+
+            flash('El Archivo solicitado no ha sido encontrado')->error()->important();
+            return back();
         }
 
         return view('sisbeca.crudNoticia.editarNoticia')->with('noticia',$noticia);
@@ -148,12 +170,24 @@ class MantenimientoNoticiaController extends Controller
 
             $noticia->url_imagen = '/images/noticias/'.$name;
 
-            $noticia->save();
+            if($noticia->save())
+            {
+                flash('Articulo Actualizado Exitosamente!','success')->important();
+            }
+            else
+            {
+                flash('Ha ocurrido un error al actualizar el articulo')->error()->important();
+            }
 
 
         }
+        else
+        {
+            flash('Debe Ingresar una imagen para la actualización del articulo')->error()->important();
+        }
 
-        return  redirect()->route('mantenimientoNoticia.index');
+    return  redirect()->route('mantenimientoNoticia.index');
+
 
 
     }
@@ -170,11 +204,21 @@ class MantenimientoNoticiaController extends Controller
         $noticia= Noticia::find($id);
         if(is_null($noticia))
         {
-            abort('404','Archivo no encontrado');
+
+            flash('El Archivo solicitado no ha sido encontrado')->error()->important();
+            return back();
         }
-        //se borra la imagen del servidor
-        unlink(public_path() .$noticia->url_imagen);
-        $noticia->delete();
+
+
+        if($noticia->delete())
+        {
+            //se borra la imagen del servidor
+            unlink(public_path() .$noticia->url_imagen);
+            flash('El Articulo ha sido Eliminado Exitosamente','info')->important();
+        }else{
+            flash('Ha Ocurrido un error al eliminar articulo')->error()->important();
+
+        }
 
         return  redirect()->route('mantenimientoNoticia.index');
     }
